requirements.txt
cryptography>=41.0.0
pyperclip>=1.8.0
textual>=0.47.0
argon2-cffi>=23.1.0
